renaming modules to match exports
