+++ src/data_table/datatables/initialize_bootstrap.js
