+++ src/data_table/datatables/initialize_col_reorder_plugin.js
